Undo & Redo
now, whenever the computer's state is changed, barring comments, the
user can undo said changes.  I will be putting in the ability to
undo/redo until last user change, system change, etc.  Will be taking
out the current ability to undo inside of single execution, so that the
system isn't left in a weird state. Because the computer itself holds
the undo info, this feature will be included in the commandline version

Will be adding more funcitonality tomorrow.

Aborts when the stack is popped on empty, may be intentional by Qt, will
address next.