Undos and Overlays
Now, the system doesn't allow for undoing partial execution.  Will
implement undos for other options as they are implemented.

Made some relatively crude attempts to port over the overlay.  Not sure
why it isn't working, but it runs without effect: most likely QHash.

There is something wrong with how the values are initialized; everything
has a value of x6565.  I think the system is reading the pointers'
values, not the value pointed to by said pointers.

Now the computer is entirely in control of the undo stack.  Will
implement levels later.