Masks and Files
Lot's of stuff:  I added the ability to mask and unmask updates to the gui.  Previously this was handeled by detaching the computer from the gui, then reataching it.  Now, when you want the system to be quiet, you put on the mask.  When you want it to speak again, you remove that mask.  Singing only occurs when the mask is off, as shown in the files.

Don't know if I want the computer to be in charge of that, but that's where it is now.

The system will try to assemble then load a file when you click File -> Open File