PreCleaner
This is a push before I go into a much more elagent and easier to maintain method of doing things.  The main difference will be that the memory views will not be handled individually, as they have been for the majority of the development.  Instead, I will be pushing those duties into a class which will allow for cleaner ties all around.