Progress Continues:PC Following, row shifting
This change introduces an ability which I think is vital for the convenient usage of this program: the insertion and deletion of rows

I plan to handle this by looking for an extensive series of 0's or, failing that, simply deleting from the last lines before the screen.  That space now free, I will move the information that is currently in the way into those slots, pushing the lot down.

A similar process will be done when deleting rows.

This would allow code to be written as if it were in a normal text editor.  I am planning on letting Ctrl-Enter insert below and Ctrl-Shift-Enter insert above.

There is much to be done to truly implement this functionallity:

Highlights hold their posiiton (or they would, I haven't hooked them up to MemWindow quite yet.)

Lars' Assemble holds its symbol table, so the mappings there will need to be updated (just the ones after the insertion line)

Another new feature that I added is the ability to track the program counter.  Simply double click the PC button and it will highlight yellow.  now, whenever the MemWindow is updated, the MemView will jump to the line where the PC rests.  Were one to want to cease this behavior, simply double clicking the PC Button will return it to its grey coloring and it will now stick where you leave it.

Currently, the program is segfaulting when malloc'd memory is accessed, which signals to me that there is something that i am overlooking.

Small Notes:

I still don't see the need for private functions if  one can simply inherit the class and make them functionally public.  I only note this because I need to implement the FollowButton so that I could get a hold of its doulble click signal.

I would propose a signifier on a function which let the coder know that this was really only meant for internal use, but which wouldn't require 2 Files and 25 lines just to find out when a user double clicks.