Steady Progress and Small adjustments
Progress is coming along nicely, but there is still much work to be done.

I have a feeling that I might need to make the code a little bit more flexible, at the cost of some elegance, but right now I am not convinced that this will be necessary.

What I am talking about is the unforeseen inability for my undoing code to handle mem_loc_t's which aren't in the computer's _memory.

I had assumed that we wouldn't be working outside of that area, so I never thought to make the code be able to handle it.

Though, as I said, I am fairly certain that I don't need to change it from the way it is now, if only because the only mem_loc_t's which are being changed outside of memory are being set up to be put into view, so there is no real reason to go back.

After much thought and planning, the ability to shift memory is approaching fruition.  All that needs be done is to complete the method for checking for errors which could be caused by shifting ( which, if I may, is almost never going to happen in the first place, but that isn't an excuse to write code which hides errors from its users ) and the code which will facilitate said shifting.