Really Small Update
I was thinking about how it would be useful to be able to go to the place where a value originated, if one didn't want to use a watch, so I put that functionality in.

There is also an attempt to highlight the line that the line the pc is pointing to, but it doesn't work.