Almost at full functional shifting
This is an intermediary push, as I don't want to lose this progress.  Right now, shifting forward works mostly, shifting backwards works mostly not.

Overall, it is coming along well.  I  consider this near core functionality, which is why I am spending so muc time on this.

I will have to set out a list of things to do to consider this ready for release, but I feel like there is not very much left.