Tiny update
I am not sure what way the stack frame coloring will be handled; I partially added a different pathway, using the changes of the R6 to signal the next color.

Up next is making the interface more reliable and more responsive.  Being able to mash the button and crash the progam is not acceptable, so I will be rewriting my code to prevent that sort of attack, exspecially when it can be expected from normal use.