Intermediate Commit
Now the undostackview has a shield on it, interrupting interaction with the view with a mask, which is then undone when the historyhandler is done setting it's index.  This has not been stress tested, so I still need to make sure everything is spick and span, but it works under good conditions.