Rudimentary connectors
There is a far better way of handling the shifting of code than I have implemented previously.

The old version requires a large search, marking when it finds a command that links to a given line.

This is incredibly inefficient and slow.

Now, I have the beginings of a system where the lines know who is pointing at them.

This will require a little more upkeep, as changing the pointing lines will demand that the pointees know, as well as other obsticales, but I believe that the effort that they require is very much so worth it.

The end goal of this is to easily implement an 'insert line' feature, where, like in a normal word document, a line can be entered without fear of breaking the rest of the program.