Saving
I looked at the minimal requirements for a shippable simulator, and I have decided that it is more important to be able to save than it is to be able to insert lines of code in the environment.

The idea is that there will be a few ways of saving: state save, where all relevant information about the ide and the contents is saved into a relatively small file; publishing, where only the area specified is saved into a portable file; and perhaps something else.