Continued Progess
Saving is coming along swimmingly.

There is a rather large hurdle which must be jumped in order to fully implement a state saving, as I plan to do.

The undos must be saved, and there is no built in method of doing that, so it is left for me to realize.

Also,  I need to be careful that everything is saved and loaded in the proper order, such that nothing is mangled.