Pretty Saves
Right now, the most important thing is to be able to save the files in an editable state.

As much as I would like to be able to say otherwise, this ide will be used as a simulator first and an ide second, most likely with a text editor off to the side.

In light of this, care should be taken to repair the code to what it was when before it was assembled.

However, I have neither the time nor the will to try to design a system for recognizing a given person's style, which would have to work on the assumption that said person's style is consistent.

Instead, I am putting in my own style as a temporary placeholder.

In the future, if there is any interest in changing this standard, it can be changed

Also, I have laid down the frameworks to allow the ide to display the data as it is expected to be.

This is partially because it is super useful, but also because it helps preserve the code that was assembled.