More Display Type Progress
I added a new display type, Color, for displaying color.

I added a menu to interact with how the data_t is changed in the tables.

Once this is done, the files will be one step closer to being savable, and then there shouldn't be much more to do to  publish.