Data Differentiation
A continuation of ground work laying, contextual data type displaying is coming closer and closer to realization.

Right now, the assembler has been equipped with the ability to note how some of the data is stored and then communicate it to the computer once assembly is complete.

This has uncovered many faults in my implementation of the mnemonic generator, so that has received some treatment as well.

STRING (without the Z) is now an option, and also gives the hint to the ide that the data therein is a set of char's.

After the reassembler is sufficiently functional, I plan to implement the ability to have the functions that set the stack be able to copy the data_t.

This is nontrivial, so will be a goal for after minimal capacity is reached.