Wasted Time
I spent no less than 3 hours trying to fix a problem that is in the standard library: fread.

fread takes in a destination, a source, a value size, and a maximum read length.

That's right.

A maximum read length.

If it hits a EOF?

Stops.

Returns how many it read.

Why it does this, when the end of file is easily identified and could be worked around, I will never know.

So thanks for that C.

Other than that, I fixed some other misspellings, as well as added some better updating.