Inch an hour; two feet a day
Fixed some hidden bugs in the assembler, as well as added some features to the context system.

It looks like all that is necessary for the minimal build is the reimplement the lc3 console ( not the control console), as well as a few other bit and bobs and fixes.

Watches are on that list, but I need to figure out how the user will add or remove them.