Amazing Grace
The display is fixed!  Don't really know what was wrong with it in the first place, what was causing those bars, but I rewrote it and it works.

Also, shifting works again.  However, it doesn't really have any ability to handle some failure, so that will be for a later date.