The only thing worse than broken code is a broken test
The console really doesn't work, but that's V2 so I am not really caring about that.

I had forgotten that I was running the computer in a separate thread when the "Next" button was pressed, allowing the "Halt" button, along with the rest of the interface, to work properly.

Saving is having some hicoughs, where the first comment is either being doubled, not be kept, or is being put in 0xFFFE ( an artifact of attempting to put it 1 before the 'starting' address.

Other than that, V1 is very close to being moved into the bug finding/fixing stage.

I will be reviewing what the minimal requirements are for that era to begin, and, once I have done that, this prject will be one step closer to being complete.

I still need to merge the two branches, by merge I mean completely overwrite.