A lack of guarantee
I have been fighting this problem for a little bit longer than I can justify, so I am going to work on input/output, saving, and cleaning up the interface.

The dream is that the saving completely repairs the code to how it was before, but I will leave those improvements to a later version.