Reluctantly Crouched at the Starting Line
Ok, this push marks the penultimate adjustment before the Beta officially begins.

This program is still buggy, the saving doesn't actually work sometimes, and it could be cleaned up a lot more than it is right now, but I have to get the ball moving.

Undoing an input crashes it, and it echos input for some reason