The Best Laid Plans
Oh, a little bit of everything in this one.

The i/o works... mostly.  Clearing the screen destroys info, so that's a thing for the later versions.

Loading can be slow, and has the opportunity to be parallelized, but that is also something for the future.

It also hangs the interface, which is very far below optimal.  I will be fixing that over the coming week, time allowing.

The I/O interface is in desperate need of a cleanup, right now the only reason that I know how to use it is because I wrote the thing.