A Brave New World
Something is very wrong with the code.

It has been wrong for so long that my computer has started to take pity on me.

It has done this in ways that I quite frankly don't understand.  However, I have the ability to switch to a new computer, so I will be doing so.