Kid:Mom, can we have unit tests? Mom: We have unit tests at home.  Unit tests at home:
While I generally agree that unit tests are very important and should be an automated part of every workflow, I am T-3 days from turning this in and every time I try to add them the automated way ( a process which requires the movement of this whole project into another project, preferably into another file as well so that everything is clean, then the creation of many, many new test projects) the thing just breaks.  So I have chosen to implement them in a way that will leave me free to fix the other things while still being sure that the code works for the most part.

The best that I can do right now is design a set of tests to show that the code will work under most circumstances and implement those inside of the code, rather than along side it.

This is strictly temporary.

Also added more watches.  I feel like 10 is enough.

SEMIFULLTEST will be rewritten to showcase the assembler.

Console will gain the ability to set breakpoints, set values, step, etc, assuming I can figure out how to get it to display into the textbox.